+++ CHANGELOG.md
@@ -21,6 +21,7 @@
   emulator can be established.
 * Fixed some minor memory leaks that occur when repeatedly loading and unloading mappers.
 * Serialize `IPokeAByteProperty` directly instead of using an intermediary anonymous object for the `PropertiesChanged` websocket event.
+* Use one contigous byte array for storing game memory, this skips a bunch of work to find the proper `ByteArray` based on address and length.
 
 # 0.8.1 (2024-11-09)
 ## Bug Fix